requirements.txt
fastapi==0.87.0
uvicorn==0.20.0
pydantic~=1.10.2
motor==3.1.1
prometheus-client==0.15.0
prometheus-fastapi-instrumentator==5.9.1
python-jose==3.3.0